# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Action Culture API
  description: |
    API backend for the EventCulture / Action Culture platform — an Algerian
    cultural heritage and events management application.
  version: 1.0.0
  contact:
    name: EventCulture Team
servers:
  - url: http://localhost:3001/api
    description: Development
  - url: https://api.actionculture.dz/api
    description: Production

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
security:
  - bearerAuth: []
  - cookieAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token sent in the Authorization header.
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: httpOnly cookie set after login / token refresh.

  # -------------------------------------------------------------------------
  # Reusable schemas
  # -------------------------------------------------------------------------
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Unauthorized
...
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            token:
              type: string
            user:
              $ref: '#/components/schemas/UserSummary'
...
    UserSummary:
      type: object
      properties:
        id_user:
          type: integer
        nom:
          type: string
        prenom:
          type: string
        email:
          type: string
        photo_url:
          type: string
          nullable: true
        type_user:
          type: integer
        statut:
          type: string
          enum: [actif, inactif, suspendu, en_attente]
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 10
        maximum: 100
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /users/refresh-token:
    post:
      tags: [Auth]
      summary: Refresh an expired access token
      description: Uses the refresh token cookie to issue a new access token.
      security: []
      responses:
        '200':
          description: New tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /dashboard/users:
    get:
      tags: [Admin]
      summary: List all users (admin)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated user list
  /dashboard/users/{id}:
    get:
      tags: [Admin]
      summary: Get user details (admin)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: User details
    put:
      tags: [Admin]
      summary: Update a user (admin)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: User updated
    delete:
      tags: [Admin]
      summary: Delete a user (admin)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: User deleted
...